Moved form to root of user settings
